requirements.txt
django==4.2.2
djangorestframework==3.14.0
django-filter==23.2
python-decouple==3.8
channels==4.0.0
channels-redis==4.1.0
channels[daphne]==4.0.0
docker==6.1.3
psycopg2-binary==2.9.6
django-debug-toolbar==4.1.0
drf_spectacular==0.26.2
drf-spectacular-sidecar==2023.6.1
djangorestframework-simplejwt==5.2.2